Add static linking for windows exe.  Remove mingw build.  Add copyright notices for fpconv to exe output.

+++ GcodeProcessorLib/fpconv.h
@@ -52,8 +52,6 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 #ifndef FPCONV_H
 #define FPCONV_H
-
-#include <stdint.h>
 /* Fast and accurate double to string conversion based on Florian Loitsch's
  * Grisu-algorithm[1].
  *
@@ -96,7 +94,7 @@ int fpconv_dtos(double fp, char dest[24], unsigned char precision);
 
 
 typedef struct Fp {
-  uint64_t frac;
+  unsigned long long frac;
   int exp;
 } Fp;
 
@@ -172,3 +170,4 @@ static Fp find_cachedpow10(int exp, int* k)
     return powers_ten[idx];
   }
 }
+
